chore(catalog): Remove kafkaless transition tables and columns
IOx has been Kafkaless for a while now and there's not any benefit for
keeping around this transitionary table and all the workarounds. This
simplifies the catalog layout. This adds a migration for both Postgres
and SQLite, although the SQLite migration is a bit painful due to lack
of support for many `ALTER TABLE` queries.