refactor: make `PartitionHashId: Copy`
The hash is 256 bits. Rustc recommends implementing `Copy` for this
size:

<https://github.com/rust-lang/rust/blob/b7bc6f88ac771e1197ac9e06bf20586eeb5d0bdf/compiler/rustc_lint/src/builtin.rs#L724-L725>

Going through an `Arc` is counterproductive here because you need to
perform pointer chasing (= cache miss) and clones need atomic operations
which are not free.

Furthermore the handling of non-`Clone` IDs is quite a PAIN as the code
simplifications in the commit demonstrate.