perf(router): no write delay backoff
Do not delay successive write attempts against upstream ingesters - a
request to A failing does not mean a request to B may fail, and sending
a request to B does not affect the load/recovery time of A - they're
independent, so delaying between rights serves no purpose.

By trying all candidates faster, router load and RAM usage is reduced as
less client connections are held open for less time (no sleeps).

This also removes the RPC_TIMEOUT bound - the RPC time remains bounded
by "rpc request timeout * number of candidate ingesters", and all
requests are aborted early if the write client disconnects. If an
ingester is offline/hanging requests, it'll quickly become unhealthy,
and further requests won't attempt to be sent to it, so there's no
"runaway delay" of all requests as successive hanging ingesters are
tried.

This also helps to ensure probe requests are always sent and recorded,
though it remains possible the client disconnects before the probe
completes.